Create Python Fundamentals
This module will first cover the BloomTech Data Science GitHub workflow and assignment submission process. It's not exactly a simple process, but you'll be doing it so often that you'll be pros in no time.  We will also review some of the Python syntax and structure you will need to know to work in a Google Colab environment.