Hypothesis Testing (t-tests) & Confidence Intervals
In this module, we’re going to build on the descriptive statistics concepts we’ve already learned about as we started to explore our data. This module will introduce the idea of a “hypothesis test” and how we implement one, specifically using the t-test and t-distributions. We’ll also cover how to calculate p-values and use the results to interpret our hypothesis.

In addition, this module will cover one of the most important concepts in statistics: the Central Limit Theorem. We’ll learn about the properties of sampling distributions and how to interpret the expected mean of a sample distribution, which will, in turn, lead to the idea of confidence intervals and how we know the confidence level of our results and predictions.