Hypothesis Testing (chi-square tests)
We’ll continue with the hypothesis testing but introduce a different way to quantify the significance of our results: the chi-square test. Next, we’ll cover when to use a chi-square test and some of the typical applications. Finally, to expand on the previous module, we’ll look at how to determine the p-value using a chi-square test and how to use it to interpret our null hypothesis.