Bayesian Statistics
So far, all statistics we’ve studied are frequentist statistics - drawing conclusions about an entire population or process by generalizing from the frequency or proportion of observed data. Frequentist statistics is well established and is still the default in most situations - but Bayesian statistics is an increasingly popular approach and an arguably better model for our brains and how we form and update beliefs based on evidence.