Simple Linear Correlation and Regression
In this module, we are going to learn how to measure the linear relationship between quantitative variables. We'll do this by creating scatter plots using the Seaborn library and calculating the linear correlation to measure the strength and direction of the relationship. Finally, we'll model the relationship between the two variables with simple linear regression.